OutputFile()を別ファイルに分けた
