+++ pg02/main.go
@@ -6,21 +6,6 @@ import (
 	"io"
 )
 
-func OutputFile(filename string) {
-	fp, err := os.Open(filename)
-	if err != nil {
-		fmt.Fprintln(os.Stderr, "File Open failed")
-		return
-	}
-	defer fp.Close()
-
-	_, err = io.Copy(os.Stdout, fp)
-	if err != nil {
-		fmt.Fprintln(os.Stderr, "File Copy failed")
-		return
-	}
-}
-
 func main() {
 	if len(os.Args) > 2 || len(os.Args) <= 1 {
 		fmt.Fprintln(os.Stderr, "Invalid argument...")
